add nuget.Dockerfile to support creating nuget package; move interfaces classes into LanceDbClient; add support for devcontainer
