+++ .vscode/launch.json
@@ -1,0 +1,36 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+          // Use IntelliSense to find out which attributes exist for C# debugging
+          // Use hover for the description of the existing attributes
+          // For further information visit https://github.com/OmniSharp/omnisharp-vscode/blob/master/debugger-launchjson.md
+          "name": ".NET Launch ApiTestbed",
+          "type": "coreclr",
+          "request": "launch",
+          //"preLaunchTask": "build",
+          // If you have changed target frameworks, make sure to update the program path.
+          "program": "${workspaceFolder}/cs/hello_ffi/ApiTestbed/bin/Debug/net8.0/ApiTestbed.dll",
+          "args": [],
+          "cwd": "${workspaceFolder}",
+          "stopAtEntry": false,
+          "env": {
+              "ASPNETCORE_ENVIRONMENT": "Development"
+          },
+          "console":"integratedTerminal",
+          "justMyCode": true,
+          "sourceFileMap": {
+              "/Views": "${workspaceFolder}/Views"
+          }
+      },
+      {
+          "name": ".NET Attach",
+          "type": "coreclr",
+          "request": "attach",
+          "processId": 1545
+      }        
+    ]
+}
