+++ cs/hello_ffi/LanceDbClient/BadVectorHandling.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public enum BadVectorHandling
     {
