+++ cs/hello_ffi/LanceDbClient/CompactionMetrics.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface;
+﻿namespace LanceDbClient;
 
 public class CompactionMetrics
 {
