+++ cs/hello_ffi/LanceDbClient/Ffi.cs
@@ -8,9 +8,9 @@ namespace LanceDbClient;
 static partial class Ffi
 {
     // Use this one for local builds
-    private const string DllName = "../../../../../../rust/target/debug/liblance_sync_client.so";
+    //private const string DllName = "../../../../../../rust/target/debug/liblance_sync_client.so";
     // Use this one for Docker 
-    //private const string DllName = "liblance_sync_client.so";
+    private const string DllName = "liblance_sync_client.so";
     
     /* Delegate types */
