+++ cs/hello_ffi/LanceDbClient/ILanceDb.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public interface ILanceDb
     {
