+++ cs/hello_ffi/LanceDbClient/ILanceHybridQueryBuilder.cs
@@ -1,8 +1,8 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public interface ILanceHybridQueryBuilder : ILanceQueryBuilder
     {
-        ILanceHybridQueryBuilder Metric(Metric metric = LanceDbInterface.Metric.L2);
+        ILanceHybridQueryBuilder Metric(Metric metric = LanceDbClient.Metric.L2);
         ILanceHybridQueryBuilder NProbes(int nProbes);
         ILanceHybridQueryBuilder RefineFactor(int refineFactor);
     }
