+++ cs/hello_ffi/LanceDbClient/ILanceMergeInsertBuilder.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public interface ILanceMergeInsertBuilder
     {
