+++ cs/hello_ffi/LanceDbClient/ILanceVectorQueryBuilder.cs
@@ -1,8 +1,8 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public interface ILanceVectorQueryBuilder : ILanceQueryBuilder
     {
-        ILanceVectorQueryBuilder Metric(Metric metric = LanceDbInterface.Metric.L2);
+        ILanceVectorQueryBuilder Metric(Metric metric = LanceDbClient.Metric.L2);
         ILanceVectorQueryBuilder NProbes(int nProbes);
         ILanceVectorQueryBuilder RefineFactor(int refineFactor);
     }
