+++ cs/hello_ffi/LanceDbClient/IndexType.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public enum IndexType
     {
