+++ cs/hello_ffi/LanceDbClient/MergeInsertBuilderAsync.cs
@@ -1,5 +1,4 @@
 using Apache.Arrow;
-using LanceDbInterface;
 
 namespace LanceDbClient;
