+++ cs/hello_ffi/LanceDbClient/Metric.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public enum Metric
     {
