+++ cs/hello_ffi/LanceDbClient/OptimizeStats.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface;
+﻿namespace LanceDbClient;
 
 public class OptimizeStats
 {
