+++ cs/hello_ffi/LanceDbClient/RRFReranker.cs
@@ -1,4 +1,3 @@
-using LanceDbInterface;
 
 namespace LanceDbClient;
