+++ cs/hello_ffi/LanceDbClient/ScalarIndexType.cs
@@ -1,4 +1,4 @@
-﻿namespace LanceDbInterface
+﻿namespace LanceDbClient
 {
     public enum ScalarIndexType
     {
