+++ cs/hello_ffi/LanceDbClientTests/Tables.cs
@@ -1,6 +1,5 @@
 using Apache.Arrow;
 using LanceDbClient;
-using LanceDbInterface;
 
 namespace LanceDbClientTests;
