--- cs/hello_ffi/LanceDbInterface/LanceDbInterface.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Apache.Arrow" Version="17.0.0" />
-      <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
-    </ItemGroup>
-
-</Project>
