+++ cs/hello_ffi/hello_ffi.sln
@@ -1,7 +1,5 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LanceDbInterface", "LanceDbInterface\LanceDbInterface.csproj", "{2958658E-E59E-4EA8-B21E-A159D603E5C1}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LanceDbClient", "LanceDbClient\LanceDbClient.csproj", "{53137863-AAA3-4151-9A4A-8E1CBAD1B4CC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiTestbed", "ApiTestbed\ApiTestbed.csproj", "{782DCCAE-3886-4972-90BE-8B6BA0786B0B}"
@@ -14,10 +12,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2958658E-E59E-4EA8-B21E-A159D603E5C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2958658E-E59E-4EA8-B21E-A159D603E5C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2958658E-E59E-4EA8-B21E-A159D603E5C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2958658E-E59E-4EA8-B21E-A159D603E5C1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{53137863-AAA3-4151-9A4A-8E1CBAD1B4CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{53137863-AAA3-4151-9A4A-8E1CBAD1B4CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{53137863-AAA3-4151-9A4A-8E1CBAD1B4CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
