+++ cs/hello_ffi/hello_ffi.sln.DotSettings.user
@@ -33,22 +33,22 @@
 	<s:String x:Key="/Default/Environment/AssemblyExplorer/XmlDocument/@EntryValue">&lt;AssemblyExplorer&gt;
   &lt;Assembly Path="/home/herbert/.nuget/packages/apache.arrow/17.0.0/lib/net8.0/Apache.Arrow.dll" /&gt;
 &lt;/AssemblyExplorer&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=0e8a6b1d_002Dbcfa_002D4e9f_002D9c6c_002Debfce37508a8/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="SimpleMergeInsert" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;Project Location="/home/herbert/Rust/lancedb_csharp/cs/hello_ffi/LanceDbClientTests" Presentation="&amp;lt;LanceDbClientTests&amp;gt;" /&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=0e8a6b1d_002Dbcfa_002D4e9f_002D9c6c_002Debfce37508a8/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="SimpleMergeInsert" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;Project Location="\home\herbert\Rust\lancedb_csharp\cs\hello_ffi\LanceDbClientTests" Presentation="&amp;lt;LanceDbClientTests&amp;gt;" /&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=324bbfbc_002D55e7_002D448b_002Da7d5_002Db439c39cbf4c/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="FullTextSearchIndexStats" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;Project Location="/home/herbert/Rust/lancedb_csharp/cs/hello_ffi/LanceDbClientTests" Presentation="&amp;lt;LanceDbClientTests&amp;gt;" /&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=324bbfbc_002D55e7_002D448b_002Da7d5_002Db439c39cbf4c/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="FullTextSearchIndexStats" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;Project Location="\home\herbert\Rust\lancedb_csharp\cs\hello_ffi\LanceDbClientTests" Presentation="&amp;lt;LanceDbClientTests&amp;gt;" /&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=9809afec_002Df963_002D4f2f_002Dbd6a_002Deced8b9dc472/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="TestOptimizeRowsAsync" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;TestAncestor&gt;
-    &lt;TestId&gt;NUnit3x::8EBD6D2C-11FC-44D9-9DB4-ED9095674B7E::net8.0::LanceDbClientTests.Tests.TestOptimizeRowsAsync&lt;/TestId&gt;
-  &lt;/TestAncestor&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=9809afec_002Df963_002D4f2f_002Dbd6a_002Deced8b9dc472/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="TestOptimizeRowsAsync" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;TestAncestor&gt;&#xD;
+    &lt;TestId&gt;NUnit3x::8EBD6D2C-11FC-44D9-9DB4-ED9095674B7E::net8.0::LanceDbClientTests.Tests.TestOptimizeRowsAsync&lt;/TestId&gt;&#xD;
+  &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=bd32ed6b_002Deb14_002D4aa0_002D910c_002D8935f1fc3840/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="Tests" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;TestAncestor&gt;
-    &lt;TestId&gt;NUnit3x::8EBD6D2C-11FC-44D9-9DB4-ED9095674B7E::net8.0::LanceDbClientTests.Tests&lt;/TestId&gt;
-  &lt;/TestAncestor&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=bd32ed6b_002Deb14_002D4aa0_002D910c_002D8935f1fc3840/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" Name="Tests" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;TestAncestor&gt;&#xD;
+    &lt;TestId&gt;NUnit3x::8EBD6D2C-11FC-44D9-9DB4-ED9095674B7E::net8.0::LanceDbClientTests.Tests&lt;/TestId&gt;&#xD;
+  &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=bfdd8fd3_002D3b6b_002D442b_002D9fce_002D3f64c38aa695/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="RerankFieldOrder" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;
-  &lt;Project Location="/home/herbert/Rust/lancedb_csharp/cs/hello_ffi/LanceDbClientTests" Presentation="&amp;lt;LanceDbClientTests&amp;gt;" /&gt;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=bfdd8fd3_002D3b6b_002D442b_002D9fce_002D3f64c38aa695/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="RerankFieldOrder" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+  &lt;Project Location="\home\herbert\Rust\lancedb_csharp\cs\hello_ffi\LanceDbClientTests" Presentation="&amp;lt;LanceDbClientTests&amp;gt;" /&gt;&#xD;
 &lt;/SessionState&gt;</s:String></wpf:ResourceDictionary>
